feat: add MentorProfile view and update schema (CAAWI-34) (#26)
* feat: add MentorProfile view and update schema (CAAWI-34)

This commit introduces a new SQL view, MentorProfile, which selects
mentor-specific details from various related tables. It also includes the
addition of MentorProfile to the Prisma schema, enabling access to these details
via Prisma client. A new migration file has been created to handle the
creation of the MentorProfile view in the database.

* feat(mentors): add route to fetch mentor profiles (CAAWI-34)

A new route has been added to the mentors API. This route uses the Prisma
client to fetch all mentor profiles from the database

* feat(prisma): update mentor profile view creation syntax (CAAWI-34)

The SQL syntax for creating the MentorProfile view has been updated
to use double quotes around the view name. This change ensures
compatibility with PostgreSQL's naming conventions.

* feat(prisma): remove database clearing in seed.ts (CAAWI-34)

* feat(mentors): add authentication to GET route (CAAWI-34)

Added server session authentication to the GET route in the mentors module.
Now, it checks for a valid session before fetching mentor profiles. If the
session is not valid, it returns an 'Unauthorised' message with a 401 status.

---------

Co-authored-by: NinjaNuur <71561866+nuurcodes@users.noreply.github.com>