feat: integrate PostHog analytics (CAAWI-39) (#31)
* feat(commitlint.config.js): add 'body-max-line-length' rule (CAAWI-39)

The 'body-max-line-length' rule has been added to the commitlint config file. This rule will ensure that the body of the commit message does not exceed the specified maximum line length.

* feat(layout): integrate PostHog analytics provider (CAAWI-39)

This commit includes the integration of PostHog analytics provider into the layout. A new PostHog provider file has been added and the layout file has been updated to use this provider. The PostHog JS library has been added to the package.json and pnpm-lock.yaml files.