feat: upgrade Prisma client and add CalendlyUser model (CAAWI-22)
This commit includes an upgrade of the Prisma client from version 5.10.2 to 5.11.0. Additionally, a new model, CalendlyUser, has been added to the Prisma schema. The Profile model has been updated to include a relation to the CalendlyUser model. A new migration file has been created to reflect these changes in the database schema.