feat(auth): add calendly route for fetching access token (CAAWI-22) (#33)
* feat(auth): add calendly route for fetching access token (CAAWI-22)

* feat: implement Calendly OAuth flow (CAAWI-22)

This commit introduces a new route for connecting to Calendly via OAuth. It also modifies the existing callback route to correctly handle the OAuth response. The access and refresh tokens are not yet saved to the database, this will be implemented in a future commit.

* feat: upgrade Prisma client and add CalendlyUser model (CAAWI-22)

This commit includes an upgrade of the Prisma client from version 5.10.2 to 5.11.0. Additionally, a new model, CalendlyUser, has been added to the Prisma schema. The Profile model has been updated to include a relation to the CalendlyUser model. A new migration file has been created to reflect these changes in the database schema.

* feat: add react-calendly package to project dependencies (CAAWI-22)

* feat: implement Calendly OAuth flow (CAAWI-22)

- Added new route for Calendly callback
- Moved Calendly connect route to a new location
- Created services for fetching access token and current user from Calendly
- Updated user profile with Calendly user data
- Made Calendly user fields optional in the database
- Handled error scenarios for fetching access token

* feat: upgrade prisma and prisma-adapter dependencies (CAAWI-22)

* refactor: rename calendly service files to kebab-case (CAAWI-22)

* feat(prisma): run prisma db pull (CAAWI-41)

* feat: remove TODO comment in calendly callback route (CAAWI-22)