feat: add Calendly integration (CAAWI-48) (#38)
* feat(prisma): remove onDelete cascade for calendlyUser relation in Profile model (CAAWI-48)

* feat: add disconnect route for Calendly integration (CAAWI-48)

* feat: add Calendly integration (CAAWI-48)

This commit introduces a new feature that allows users to connect their Calendly account to the application. Users can now schedule meetings directly from the application. The integration includes a new Calendly button component and a Calendly widget. The button allows users to connect or disconnect their Calendly account, and the widget displays the user's Calendly scheduling link. The mentor profile component has also been updated to include the Calendly widget.

* feat: add user type check in ConnectCalendlyButton component (CAAWI-48)

* feat: rename ConnectCalendlyButton to CalendlyConnectionButton (CAAWI-47 & CAAWI-48)

* feat: replace InlineWidget with PopupButton in Calendly component CAAWI-48

* feat: add revoke access token functionality for Calendly API (CAAWI-48)

This commit introduces the ability to revoke access tokens for the Calendly API. This is accomplished by creating a new service that sends a POST request to the Calendly API's revoke endpoint. Additionally, modifications have been made to the disconnect route to utilize this new service. The callback route has also been updated to set cookies for the access token, refresh token, and organization.

* Update components/calendly/calendly-widget.tsx

Co-authored-by: NinjaNuur <71561866+nuurcodes@users.noreply.github.com>

* feat: add validation for Calendly environment variables (CAAWI-48)

This commit introduces a new validation schema for Calendly environment variables. It ensures that the required variables are present and are of the correct type. This schema is used in various services and routes related to Calendly API.

---------

Co-authored-by: NinjaNuur <71561866+nuurcodes@users.noreply.github.com>