feat: add unique constraints and modify database schema (CAAWI-57) (#40)
* feat: add unique constraints and modify database schema (CAAWI-57)

This commit includes the following changes:
- Added unique constraints to the `country` field in the `Location` model, `company` field in the `Occupation` model, and `scheduling_url` field in the `Profile` model in the Prisma schema.
- Renamed and modified the migration file, including the addition of a new `Gender` enum, changes to the `Profile` and `Occupation` tables, and the creation of a new `CalendlyUser` table.
- Updated the SQL commands in the migration file to reflect these changes.

* feat(seed.ts): refactor seeding process for user profiles (CAAWI-57)

* feat(seed.ts): refactor user creation in seed file (CAAWI-57)

Refactored the user creation process in the seed file to use the connectOrCreate method for location and occupation. This simplifies the code and reduces the number of database operations.

* feat: update unique constraints in Prisma schema (CAAWI-57)

- Modified unique constraints in `Location` and `Occupation` models in Prisma schema.
- Updated seed.ts to match the new unique constraints.
- Created a new migration file to reflect these changes in the database.

* feat(schema): make city, country, role, and company fields nullable (CAAWI-57)

Updated the schema for the Location and Occupation models in the Prisma schema file to make the city, country, role, and company fields nullable. Also updated the corresponding SQL migration file to reflect these changes.

* feat(seed.ts): add database clearing before seeding (CAAWI-57)

This commit adds a feature to clear the database before seeding. It also includes minor changes to the connectOrCreate function for location and occupation, making the code more explicit.