feat: add social links (CAAWI-58) (#46)
* feat: add social links (CAAWI-58)

* feat: add social links href (CAAWI-58)

* feat(users): refactor user update logic in route.ts (CAAWI-58)

Simplified the user update logic in the route.ts file by removing unnecessary condition checks for location and occupation. Replaced the upsert method with the connectOrCreate method for both location and occupation fields. This change makes the code cleaner and more efficient.

* fix: update social links validation (CAAWI-58)

* feat(users route): add conditional checks for country and role fields in user creation (CAAWI-58)

This commit introduces conditional checks for the 'country' and 'role' fields during user creation in the users route. This ensures that these fields are only processed if they contain valid data, preventing potential errors or inconsistencies in the database.

* fix: remove form dependency (CAAWI-58)

* feat: update JOIN condition in MentorProfile.sql view (CAAWI-58)

Updated the JOIN condition in the MentorProfile.sql view to match the User's id with the Profile's userId. This change ensures the correct association between User and Profile entities.

* feat(callback/route): update calendly user creation logic in callback route (CAAWI-58)

* fix: city dropdown (CAAWI-58)

---------

Co-authored-by: Abshir <abshir.ahmed@somalisintech.com>