feat: add skills model to the database (CAAWI-68) (#55)
* feat: add skills model to the database (CAAWI-68)

This commit introduces a new Skill model to the database schema and updates
the existing Profile model to include a many-to-many relationship with the
Skill model. The seed data has been updated to include a list of skills for
each user. The MentorProfile view has also been updated to aggregate and
display the skills for each mentor. Two new migration files have been added
to handle the creation of the Skill model and the update to the MentorProfile
view.

* feat: add social links and enhance skills seeding in DB (CAAWI-68)

This commit adds LinkedIn, GitHub, and BuyMeCoffee URLs to the MentorProfile view
and updates the migration.sql file accordingly. It also refactors the seed.ts
file to use upsert for skills, simplifying the code and improving efficiency.

* feat(seed.ts): change seeding restriction message and condition (CAAWI-68)

Changed the condition for seeding from 'development' to 'production' and
replaced the error throwing with a console log message and process exit.
This provides a more user-friendly message and prevents seeding in a
production environment.